+++ Kaidel/src/Platform/OpenGL/OpenGLBuffer.cpp
@@ -48,7 +48,7 @@ namespace Kaidel {
 	{
 		glBindBuffer(GL_ARRAY_BUFFER, m_RendererID);
 		if (m_MaxSize < size) {
-			m_MaxSize = size*1.5;
+			m_MaxSize = (uint64_t)((double)size*1.5);
 			glBufferData(GL_ARRAY_BUFFER, m_MaxSize, nullptr, GL_DYNAMIC_DRAW);
 		}
 		glBufferSubData(GL_ARRAY_BUFFER, 0, size, data);
